avoid assembly scanning for publisher configuration
