+++ src/NServiceBus.Core.Tests/API/APIApprovals.ApproveNServiceBus.approved.txt
@@ -2581,8 +2581,6 @@ namespace NServiceBus.Routing.MessageDrivenSubscriptions
     {
         public Publishers() { }
         public void Add(System.Type eventType, string publisher) { }
-        public void Add(System.Reflection.Assembly eventAssembly, string publisher) { }
-        public void Add(System.Reflection.Assembly eventAssembly, string eventNamespace, string publisher) { }
         public void AddByAddress(System.Type eventType, string publisherAddress) { }
         public void AddDynamic(System.Func<System.Type, NServiceBus.Routing.MessageDrivenSubscriptions.PublisherAddress> dynamicRule) { }
     }
