+++ src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
@@ -162,7 +162,7 @@
     <Compile Include="Routing\DetermineRouteForPublishBehaviorTests.cs" />
     <Compile Include="Routing\DistributionPolicyTests.cs" />
     <Compile Include="Routing\FileBasedDynamicRouting\FileRoutingTableFeatureTests.cs" />
-    <Compile Include="Routing\MessageDrivenSubscriptionsConfigExtensionsTests.cs" />
+    <Compile Include="Routing\MessageDrivenSubscriptions\MessageDrivenSubscriptionsConfigExtensionsTests.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\PublishersTests.cs" />
     <Compile Include="Routing\Routers\UnicastSendRouterConnectorTests.cs" />
     <Compile Include="Routing\EndpointInstanceTests.cs" />
